put cgralib into a separate folder for verilog gen
