+++ aha_garnet_design/conv_3_3/conv_3_3.json
@@ -418,7 +418,7 @@
           },
           "ub_hw_input_global_wrapper_stencil_op_hcompute_hw_input_global_wrapper_stencil_2_to_hw_input_global_wrapper_stencil_op_hcompute_conv_stencil_1_11":{
             "genref":"cgralib.Mem",
-            "genargs":{"has_flush":["Bool",true], "has_reset":["Bool",false], "has_stencil_valid":["Bool",true], "has_valid":["Bool",false], "num_input":["Int",1], "num_output":["Int",2], "width":["Int",16]},
+            "genargs":{"has_flush":["Bool",true], "has_reset":["Bool",false], "has_stencil_valid":["Bool",true], "has_valid":["Bool",false], "num_inputs":["Int",1], "num_outputs":["Int",2], "width":["Int",16]},
             "modargs":{"config":["Json",{"agg2sram":{"cycle_starting_addr":[4],"cycle_stride":[4,64,0],"dimensionality":3,"extent":[16,64,1],"read_data_starting_addr":[0],"read_data_stride":[0,0,0],"write_data_starting_addr":[0],"write_data_stride":[1,16,0]},"in2agg_0":{"cycle_starting_addr":[0],"cycle_stride":[1,4,64,0],"dimensionality":4,"extent":[4,16,64,1],"write_data_starting_addr":[0],"write_data_stride":[1,0,0,0]},"sram2tb":{"cycle_starting_addr":[123],"cycle_stride":[2,4,64,0],"dimensionality":4,"extent":[2,16,62,1],"mux_write_data_starting_addr":[0],"mux_write_data_stride":[1,0,0,0],"read_data_starting_addr":[0],"read_data_stride":[16,1,16,0],"write_data_starting_addr":[0],"write_data_stride":[0,1,0,0]},"stencil_valid":{"cycle_starting_addr":[130],"cycle_stride":[1,64,0],"dimensionality":3,"extent":[62,62,1]},"tb2out_0":{"cycle_starting_addr":[128],"cycle_stride":[1,4,64,0],"dimensionality":4,"extent":[4,16,62,1],"read_data_starting_addr":[0],"read_data_stride":[1,4,0,0]},"tb2out_1":{"cycle_starting_addr":[128],"cycle_stride":[1,4,64,0],"dimensionality":4,"extent":[4,16,62,1],"read_data_starting_addr":[0],"read_data_stride":[1,4,0,0]}}], "mode":["String","lake"]}
           },
           "ub_hw_input_global_wrapper_stencil_op_hcompute_hw_input_global_wrapper_stencil_2_to_hw_input_global_wrapper_stencil_op_hcompute_conv_stencil_1_11_clk_en_const":{
