+++ cgralib.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "coreir/common-macros.h"
+#include "coreir-c/ctypes.h"
+
+#ifdef __cplusplus
+#include "coreir.h"
+COREIR_GEN_CPP_API_DECLARATION_FOR_LIBRARY(cgralib);
+#endif
+
+COREIR_GEN_C_API_DECLARATION_FOR_LIBRARY(cgralib);
